# Optimized decoder and paved the way for parsing extensions.