… upb/msg_internal.h
… upbc/protoc-gen-upb.cc