The primary motivation for this change is to avoid referring to the `upb_msglayout` object when we are trying to fetch the `upb_msglayout` object for a sub-message. This will help pave the way for parsing extensions. We also implement several optimizations so that we can make this change without regressing performance.

Normally we compute the layout for a sub-message field like so:

```c
const upb_msglayout *get_submsg_layout(
    const upb_msglayout *layout,
    const upb_msglayout_field *field) {
  return layout->submsgs[field->submsg_index]
}
```

The reason for this indirection is to avoid storing a pointer directly in `upb_msglayout_field`, as this would double its size (from 12 to 24 bytes on 64-bit architectures) which is wasteful as this pointer is only needed for message typed fields.

However `get_submsg_layout` as written above does not work for extensions, as they will not have entries in the message's `layout->submsgs` array by nature, and we want to avoid creating an entire fake `upb_msglayout` for each such extension since that would also be wasteful.

This change removes the dependency on `upb_msglayout` by passing down the `submsgs` array instead:

```c
const upb_msglayout *get_submsg_layout(
    const upb_msglayout *const *submsgs,
    const upb_msglayout_field *field) {
  return submsgs[field->submsg_index]
}
```

This will pave the way for parsing extensions, as we can more easily create an alternative `submsgs` array for extension fields without extra overhead or waste.

Along the way several optimizations presented themselves that allow a nice increase in performance:

1. Passing the parsed `wireval` by address instead of by value ended up avoiding an expensive and useless stack copy (this is on Clang, which was used for all measurements).
2. When field numbers are densely packed, we can find a field by number with a single indexed lookup instead of linear search. At codegen time we can compute the maximum field number that will allow such an indexed lookup.
3. For fields that do require linear search, we can start the linear search at the location where we found the previous field, taking advantage of the fact that field numbers are generally increasing.
4. When the hasbit index is less than 32 (the common case) we can use a less expensive code sequence to set it.
5. We check for the hasbit case before the oneof case, as optional fields are more common than oneof fields.

Benchmark results indicate a 20-25% reduction in parse time with a small code size increase:

```
name                                      old time/op  new time/op  delta
ArenaOneAlloc                             21.3ns ± 0%  21.9ns ± 1%   +3.08%  (p=0.000 n=11+12)
ArenaInitialBlockOneAlloc                 6.32ns ± 0%  6.32ns ± 0%     ~     (p=0.219 n=12+12)
LoadDescriptor_Upb                        53.5µs ± 1%  51.2µs ± 2%   -4.30%  (p=0.000 n=12+12)
LoadAdsDescriptor_Upb                     2.75ms ± 0%  2.67ms ± 2%   -2.68%  (p=0.000 n=10+10)
LoadDescriptor_Proto2                      240µs ± 0%   239µs ± 0%   -0.22%  (p=0.000 n=12+12)
LoadAdsDescriptor_Proto2                  12.6ms ± 0%  12.8ms ± 0%   +1.17%  (p=0.000 n=12+11)
Parse_Upb_FileDesc<UseArena,Copy>         13.2µs ± 0%  10.1µs ± 0%  -23.85%  (p=0.000 n=12+12)
Parse_Upb_FileDesc<UseArena,Alias>        11.5µs ± 0%   8.8µs ± 0%  -22.92%  (p=0.000 n=11+11)
Parse_Upb_FileDesc<InitBlock,Copy>        12.7µs ± 0%   9.6µs ± 0%  -24.68%  (p=0.000 n=11+12)
Parse_Upb_FileDesc<InitBlock,Alias>       11.0µs ± 0%   8.5µs ± 0%  -22.89%  (p=0.000 n=12+11)
Parse_Proto2<FileDesc,NoArena,Copy>       29.3µs ± 0%  29.4µs ± 0%   +0.30%  (p=0.001 n=12+12)
Parse_Proto2<FileDesc,UseArena,Copy>      20.4µs ± 2%  21.1µs ± 2%   +3.38%  (p=0.000 n=12+12)
Parse_Proto2<FileDesc,InitBlock,Copy>     16.5µs ± 0%  17.5µs ± 0%   +5.98%  (p=0.000 n=12+12)
Parse_Proto2<FileDescSV,InitBlock,Alias>  16.2µs ± 0%  17.4µs ± 0%   +7.52%  (p=0.000 n=12+11)
SerializeDescriptor_Proto2                5.35µs ± 1%  5.47µs ± 1%   +2.19%  (p=0.000 n=12+12)
SerializeDescriptor_Upb                   12.9µs ± 0%  13.1µs ± 0%   +1.41%  (p=0.000 n=12+11)


    FILE SIZE        VM SIZE    
 --------------  -------------- 
  +0.6%     +64  +0.6%     +64    upb/decode.c
    +0.7%     +64  +0.7%     +64    decode_msg
  +0.3%     +48  +0.3%     +48    upb/def.c
    +1.1%     +48  +1.1%     +48    _upb_symtab_addfile
  -3.6%    -112  [ = ]       0    [Unmapped]
  [ = ]       0  +0.1%    +112    TOTAL
```